+++ src/globals.h
@@ -14,6 +14,7 @@
 #include "signAddAnonymityRevoker.h"
 #include "signAddBakerOrUpdateBakerKeys.h"
 #include "signAddIdentityProvider.h"
+#include "signConfigureBaker.h"
 #include "signConfigureDelegation.h"
 #include "signCredentialDeployment.h"
 #include "signEncryptedAmountTransfer.h"
@@ -65,6 +66,7 @@ typedef enum {
     TRANSFER_WITH_MEMO = 22,
     ENCRYPTED_AMOUNT_TRANSFER_WITH_MEMO = 23,
     TRANSFER_WITH_SCHEDULE_WITH_MEMO = 24,
+    CONFIGURE_BAKER = 25,
     CONFIGURE_DELEGATION = 26
 } transactionKind_e;
 
@@ -157,6 +159,7 @@ typedef union {
     signAddBakerContext_t signAddBaker;
     signUpdateBakerStakeContext_t signUpdateBakerStake;
     signUpdateBakerRestakeEarningsContext_t signUpdateBakerRestakeEarnings;
+    signConfigureBaker_t signConfigureBaker;
     signConfigureDelegationContext_t signConfigureDelegation;
 
     signExchangeRateContext_t signExchangeRateContext;