+++ src/signUpdateTransactionFeeDistribution.c
@@ -32,17 +32,14 @@ void handleSignUpdateTransactionFeeDistribution(uint8_t *cdata, volatile unsigne
 
     // Baker fee is first 4 bytes
     uint32_t bakerFee = U4BE(cdata, 0);
-    int bakerFeeLength = numberToText(ctx->baker, sizeof(ctx->baker), bakerFee);
+    fractionToText(bakerFee, ctx->baker, sizeof(ctx->baker));
     cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, 4, NULL, 0);
     cdata += 4;
-    uint8_t fraction[8] = "/100000";
-    memmove(ctx->baker + bakerFeeLength, fraction, 8);
 
     // Gas account fee is the next 4 bytes
     uint32_t gasAccountFee = U4BE(cdata, 0);
-    int gasAccountFeeLength = numberToText(ctx->gasAccount, sizeof(ctx->gasAccount), gasAccountFee);
+    fractionToText(gasAccountFee, ctx->gasAccount, sizeof(ctx->gasAccount));
     cx_hash((cx_hash_t *) &tx_state->hash, 0, cdata, 4, NULL, 0);
-    memmove(ctx->gasAccount + gasAccountFeeLength, fraction, 8);
 
     ux_flow_init(0, ux_sign_transaction_dist, NULL);
     *flags |= IO_ASYNCH_REPLY;